refactor(app,ui): :recycle: extract table to separate component
