+++ src/components/index.ts
@@ -1,5 +1,6 @@
 export * from './ClassDetail'
 export * from './ClassStatusSwitch'
+export * from './ClassTable'
 export * from './HeaderBar'
 export * from './SideNav'
 export * from './TagGroup'
